spring security com rest ok
e o controle do form pelo spring security ainda não foi habilitado

+++ src/main/java/ngdemo/repositories/contract/Repository.java
@@ -9,5 +9,5 @@ public interface Repository<T, PK extends Serializable> {
 
     T getById(PK id);
 
-    int getCount();
+    Long getCount();
 }
